+++ src/main/java/fi/dy/masa/malilib/MaLiLibReference.java
@@ -2,6 +2,7 @@ package fi.dy.masa.malilib;
 
 public class MaLiLibReference
 {
-    public static final String MOD_ID = "mafglib";
-    public static final String MOD_NAME = "MaLiLib";
+    public static final String MOD_ID = "malilib";
+    public static final String NAMESPACE = "malilib";
+    public static final String MOD_NAME = "MaFgLib";
 }