+++ src/main/java/fi/dy/masa/malilib/MaLiLibReference.java
@@ -3,5 +3,6 @@ package fi.dy.masa.malilib;
 public class MaLiLibReference
 {
     public static final String MOD_ID = "mafglib";
+    public static final String NAMESPACE = "malilib";
     public static final String MOD_NAME = "MaFgLib";
 }