+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Changelog
+- change modid to `mafglib`
 - sync malilib 0.16.2
 - auto upload