+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
+- change modid to `mafglib`
 - auto upload