+++ CHANGELOG.md
@@ -1,4 +1,2 @@
 ## Changelog
-- change modid to `mafglib`
-- sync malilib 0.16.2
-- auto upload
+- fix icon missing issue