+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## Changelog
-- change modid to `mafglib`
-- auto upload
+- fix icon missing issue
+- Deprecated `ForgePlatformCompat`, all function move to `ForgePlatformUtils`
+- Add old MaFgLib modid compat
+- Update Architectury Loom to 1.4